pulling orders

+++ mystery-box/src/views/register.jsx
@@ -13,6 +13,7 @@ class Register extends Component {
     email: '',
     password: '',
     error: '',
+    confirmPassword: '',
   };
 
   handleChange = ({ target }) => {
@@ -23,48 +24,53 @@ class Register extends Component {
 
   handleSubmit = (event) => {
     event.preventDefault();
-    const { username, email, password } = this.state;
+    const { username, email, password, confirmPassword } = this.state;
     const { updateUser } = this.props;
 
     const credentials = {
       email,
       password,
-      username
+      username,
     }
 
-    this.setState({
-      error: ''
-    }, async () => {
+    if (password === confirmPassword) {
 
-      try {
-        const result = await Register.service.login(credentials);
-        console.log(result);
+      this.setState({
+        error: ''
+      }, async () => {
 
-        if (!result.success) {
-          const errors = Object.values(result.errors).join(' ');
-          throw new Error(errors);
+        try {
+          const result = await Register.service.login(credentials);
+          console.log(result);
+
+          if (!result.success) {
+            const errors = Object.values(result.errors).join(' ');
+            throw new Error(errors);
+          }
+
+          const autoLogin = await Register.login.login(credentials);
+          console.log(autoLogin);
+
+          window.localStorage.setItem('auth_token', autoLogin.token);
+
+          if (autoLogin.success) {
+
+            updateUser({
+              isLoggedin: true,
+              updateUser,
+              ...autoLogin.user
+            });
+          }
+
+        } catch (error) {
+          this.setState({
+            error: error.message,
+          })
         }
-
-        const autoLogin = await Register.login.login(credentials);
-        console.log(autoLogin);
-
-        window.localStorage.setItem('auth_token', autoLogin.token);
-
-        if (autoLogin.success) {
-
-          updateUser({
-            isLoggedin: true,
-            updateUser,
-            ...autoLogin.user
-        });
-        }
-
-      } catch (error) {
-        this.setState({
-          error: error.message,
-        })
-      }
-    })
+      })
+    } else {
+      console.log('Password does not match!')
+    }
   }
 
   render() {
@@ -83,13 +89,13 @@ class Register extends Component {
       <div className="form-wrapper">
         {
           error.length
-            ? <div>Something went wrong: {error}</div>
+            ? <div className="errors">Something went wrong: {error}</div>
             : null
         }
         <h1>Register</h1>
         <form onSubmit={this.handleSubmit}>
           <div className="form-group">
-            <label htmlFor="email">E-mail</label>
+            {/* <label htmlFor="email">E-mail</label> */}
             <input
               type="text"
               name="email"
@@ -98,7 +104,7 @@ class Register extends Component {
               value={email}
               onChange={this.handleChange} /></div>
           <div className="form-group">
-            <label htmlFor="username">Username</label>
+            {/* <label htmlFor="username">Username</label> */}
             <input
               type="text"
               name="username"
@@ -107,7 +113,7 @@ class Register extends Component {
               value={username}
               onChange={this.handleChange} /></div>
           <div className="form-group">
-            <label htmlFor="password">Password</label>
+            {/* <label htmlFor="password">Password</label> */}
             <input
               type="password"
               name="password"
@@ -116,11 +122,11 @@ class Register extends Component {
               value={password}
               onChange={this.handleChange} /></div>
           <div className="form-group">
-            <label htmlFor="confirmPassword">Confirm Password</label>
+            {/* <label htmlFor="confirmPassword">Confirm Password</label> */}
             <input
               type="password" name="confirmPassword"
-              id="confirmPassword" placeholder="Enter your password again" value={confirmPassword} /></div>
-          <input type="submit" value="Register" />
+              id="confirmPassword" placeholder="Confirm your password" value={confirmPassword} onChange={this.handleChange}/></div>
+          <button type="submit" className="btn btn-primary">Register</button>
         </form>
       </div>
     )
